+++ backend/DatabaseScripts/SqlScripts/0037_AlterTable_graphql_payday_status_add_protocol_version.sql
@@ -1,0 +1,1 @@
+﻿ALTER TABLE graphql_payday_status ADD COLUMN protocol_version INT NOT NULL DEFAULT 0;